+++ instrumentation/rxjava/rxjava-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/rxjava2/RxJava2InstrumentationModule.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.rxjava3;
+package io.opentelemetry.javaagent.instrumentation.rxjava2;
 
 import com.google.auto.service.AutoService;
 import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
@@ -12,10 +12,10 @@ import java.util.Collections;
 import java.util.List;
 
 @AutoService(InstrumentationModule.class)
-public class RxJava3InstrumentationModule extends InstrumentationModule {
+public class RxJava2InstrumentationModule extends InstrumentationModule {
 
-  public RxJava3InstrumentationModule() {
-    super("rxjava3");
+  public RxJava2InstrumentationModule() {
+    super("rxjava2");
   }
 
   @Override
